Fix Flutter web compilation error
